Extract the word mappings screen into a separate class
